+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added support for the HTML editor.
 * Added support for the CSS editor.
 * Fixed annotation colouring in the Java editor.
+* Added support for the JavaScript editor.
 
 0.2.1 (2011-01-07)
 ==================